backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
python-dotenv>=1.0.0
httpx>=0.25.0 # Added for async HTTP requests
# Add database drivers later if needed (e.g., psycopg2-binary, motor)